fix: limit unsigned integer range. #1151
[summary]

tinyint: signed range is -128 to 127, unsigned range is 0 to 127.

smallint: signed range is -32768 to 32767, unsigned range is 0 to 32767.

mediumint: signed range is -8388608 to 8388607, unsigned range is 0 to 8388607.

int/integer: signed range is -2147483647 to 2147483647, unsigned range is 0 to 2147483647.
Note: -2147483648 is not allowed in tianmu engine currently.

bigint: signed range is -9223372036854775806 to 9223372036854775807, unsigned range is 0 to 9223372036854775807.
Note: -9223372036854775807 will be set to null and -9223372036854775808 will ret out of range.